Lab9/requirement.txt
ansicon==1.89.0
appdirs==1.4.4
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-admin-logs==1.0.1
django-allauth==0.44.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-mysql==3.10.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-picklefield==3.0.1
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jinxed==1.0.1
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.4
phonenumbers==8.12.14
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.1.0
wcwidth==0.2.5
